refactor: move Migration to separate class
